requirements.txt
setuptools>=68.0.0,~=68.0
python-dateutil>=2.8.2,~=2.8
pytz>=2023.3,<2024
numpy>=1.25.0,~=1.25
pandas>=2.0.3,~=2.0
scipy>=1.11.1,~=1.11
tables>=3.8.0,~=3.8
openpyxl>=3.0.10,~=3.0
matplotlib>=3.7.1,~=3.7
seaborn>=0.12.2,~=0.12
corsys>=0.8.4,~=0.8
corsys>=0.2.10,~=0.2
